as_poly: raise NotImplementedError instead of returning None
Aaron commented on this:

"[...] Also, I thought polys12 had a commit making that
NotImplementedError.  Maybe it was in my branch.

Well, I didn't find that commit in either place, but it needs to be
made. It's also the problem in issues 2068 and 1732. I think making that
raise NotImplementedError when it can't make a poly might fix at least
the first of those."

return None instead of NotImplementedError