limit sorting of quadratic and binomial roots to ZZ
The sorting that is being done in roots_binomial and roots_quadratic
is only being done to make sure that the RootOf instances for these
be correlated with the returned values. When the domain is ZZ then
the various tests of signed quantities is guranteed to not fail since
they involve determining the sign of a Rational. For anything more
complicated, there is no need to sort at this private function level
so we resist the temptation to do so even if we can (e.g. the tests
added seem simple enough and the roots *could* be sorted but there
is no need to do so and no guarantee that the assumptions system will
always be able to handle more complicated numerical expression).