improve solveset radical solving
Some clarication notes were made to unrad and a loop
was added to make sure that the returned expression is
fully expanded. A test requiring this was added.

_solve_radical was modified to use RootOf instances when
possible to speed up the solution finding process.

_invert_real now only takes real_root when the exponent
is an Integer

solveset_[real|complex] now are aware of an expression that
is a Pow when searching for radicals. tests were added
or unXFAILed that depended on this modification.