feat(polys): Add DomainMatrix based on poly elements
Modify tests using units and prefixes with linsolve. The results from
linsolve are correct but in a different form because prefixes like kilo
were previously converted to Rational like 1000 in linsolve. Changing
the order of arithmetic by using DomainMatrix has affected this because
multiplication of prefixes like kilo is not associative:

Related issue:
https://github.com/sympy/sympy/issues/18988