sort match terms; ignore bound
When matching, args are ordered but the default sort key
prioritizes symbols over leading Numbers on Mul. This means
5*x < 2*y. By sorting by the leading Number (as the added
test shows) a better match can be made. Previously the test
gave

>>> (5*y-x).match(5*Wild('x')-Wild('y'))
{Y_: -5*y, X_: -x/5}

Now it gives

>>> (5*y-x).match(5*Wild('x')-Wild('y'))
{Y_: x, X_: y}

match ignores bound symbols unless requested

This means that when matching expressions with
potentially different bound symbols, if you
aren't interested in the identity of the
bound symbols you can leave them out of the
pattern.