fix(core): Keep identities when evaluate=False
Change str and pretty printers when handling unevaluated Muls. If a Mul
has two or more Numbers in its args or if the first arg is not a Number
then it must be an unevaluated Mul. In this case we print the result
strictly in order of args with no attempt to group numerators and
denominators together in a fraction so that the internal unevaluated
representation becomes clear.