as_coefficients_dict accepts symbols
as_coefficients_dict accepts syms

In addition, a routine to use this to return a system of
equations to be solved for a generalized "undetermined coefficients"
situation is added and used in solve.

By doing this separation *before* passing a single equation
to `_solve` allows avoidance of having to deal with the multi-form
output of the recursive call to `solve_undetermined_coeffs`.

The latter has been tightened up to return None unless the system
is linear (as is the normal use of such systems).