add match (default False) flag to solve
It has been debated whether this is a bug fix or a
feature change. Since solving an equation as a set
of equations derived by matching coefficients on
powers of a variable is not the normal use of `solve`
it is being treated as a bug fix in this commit.

For sake of those that might be depending on this
behavior, however, the use of keyword `match=True`
will restore the behavior rather then having to
rewrite the call to `solve` with a call to `solve_undetermined_coeffs`,
e.g.

    solve(a*x + b - 2*x, a, b) to
    solve_undetermined_coeffs(a*x + b - 2*x, [a, b], x)
    instead of
    solve(a*x + b - 2*x, a, b, match=True)