complicated number now factor, not coeff
The routine tries to keep track of all numbers
in the coeff, but when there is a complicated
number (like an integral that has a numerical
answer) trying to evaluate it to store it in
the coeff is expensive. So now it is kept with
the other factors. In addition, complex is not
used to evaluate any of the numbers that are
known to be real or not. They are kept in
sympy form. I'm not sure if this will cause
issues or not since it is easier to keep track
of floats.